// src/app/model/Abilita.ts
export class Abilita {
    allettare: number;
    arteDellaGuerra: number;
    atletica: number;
    cavalcare: number;
    convincere: number;
    empatia: number;
    esibirsi: number;
    furto: number;
    intimidire: number;
    istruzione: number;
    mira: number;
    mischia: number;
    nascondersi: number;
    navigazione: number;
    notare: number;
    rissa: number;
}